Rename gateways to storage
#time 11m
