--- omigami/spectra_matching/spec2vec/gateways/__init__.py
@@ -1,3 +1,0 @@
-from omigami.spectra_matching.spec2vec.gateways.fs_document_iterator import (
-    FileSystemDocumentIterator,
-)
