+++ omigami/spectra_matching/spec2vec/storage/__init__.py
@@ -1,0 +1,3 @@
+from omigami.spectra_matching.spec2vec.storage.fs_document_iterator import (
+    FileSystemDocumentIterator,
+)
