+++ omigami/spectra_matching/spec2vec/storage/fs_document_iterator.py
