+++ omigami/spectra_matching/spec2vec/storage/spectrum_document.py
