+++ omigami/spectra_matching/spec2vec/tasks/process_spectrum/task.py
@@ -3,19 +3,19 @@ from typing import Set, List
 
 from prefect import Task
 
-from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.spectra_matching.spec2vec.entities.spectrum_document import (
     SpectrumDocumentData,
-)
-from omigami.spectra_matching.spec2vec.gateways.spectrum_document import (
-    SpectrumDocumentDataGateway,
 )
 from omigami.spectra_matching.spec2vec.helper_classes.progress_logger import (
     TaskProgressLogger,
 )
+from omigami.spectra_matching.spec2vec.storage.spectrum_document import (
+    SpectrumDocumentDataGateway,
+)
 from omigami.spectra_matching.spec2vec.tasks.process_spectrum.spectrum_processor import (
     SpectrumProcessor,
 )
+from omigami.spectra_matching.storage import RedisSpectrumDataGateway, FSDataGateway
 from omigami.utils import merge_prefect_task_configs
