+++ omigami/test/spectra_matching/spec2vec/gateways/test_redis_document_dgw.py
@@ -6,7 +6,7 @@ from spec2vec import SpectrumDocument
 from omigami.config import SPECTRUM_ID_PRECURSOR_MZ_SORTED_SET
 from omigami.spectra_matching.spec2vec.config import PROJECT_NAME
 from omigami.spectra_matching.spec2vec.entities.embedding import Spec2VecEmbedding
-from omigami.spectra_matching.spec2vec.gateways.redis_spectrum_document import (
+from omigami.spectra_matching.spec2vec.storage.redis_spectrum_document import (
     RedisSpectrumDocumentDataGateway,
 )
 from omigami.spectra_matching.storage import RedisSpectrumDataGateway
